Working on video loader. Got some crappy test data and tried things out.  Bundler seems to be very hit-and-miss.  Some frames get a correct matrix, while most get nothing at all.  Not sure what to do about this....
Also, split up loading code so it is possible to load a temporary file from bundler.  This makes it easier to examine the data and figure where exactly things went wrong.

Volume cuts is a mess and probably doesn't work.  May end up doing a rewrite of the code later.  Need to think about it more carefully.